Rename package resource to resources.
Fixes #259.
