+++ api/src/main/java/io/opentelemetry/resources/package-info.java
@@ -17,7 +17,7 @@
 /**
  * API for resource information population.
  *
- * <p>The resource library primarily defines a type "Resource" that captures information about the
+ * <p>The resources library primarily defines a type "Resource" that captures information about the
  * entity for which stats or traces are recorded. For example, metrics exposed by a Kubernetes
  * container can be linked to a resource that specifies the cluster, namespace, pod, and container
  * name.
@@ -34,4 +34,4 @@
  * zero. They SHOULD start with a domain name and separate hierarchies with / characters, e.g.
  * k8s.io/namespace/name.
  */
-package io.opentelemetry.resource;
+package io.opentelemetry.resources;
